fix: definition of circle now uses radius unsigned
We cannot have circles of negative radius

+++ include/draw.h
@@ -9,7 +9,7 @@ struct Point{
 
 bool is_valid(Image &img, Point p);
 void draw_rectangle(Image &img, Point p1, size_t w, size_t h, Color col);
-void draw_circle(Image &img, Point c, int radiu, Color col);
+void draw_circle(Image &img, Point c, size_t radiu, Color col);
 void draw_line_2P(Image &img, Point p1, Point p2, Color col);
 void draw_triangle(Image &img, Point p1, Point p2, Point p3);
 #endif
